requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.0.3
Django==2.0.3
djangorestframework==3.7.7
feedfinder2==0.0.4
feedparser==5.2.1
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
idna==2.6
itypes==1.1.0
jieba3k==0.35.1
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
newspaper3k==0.2.6
nltk==3.2.5
Pillow==5.0.0
python-dateutil==2.7.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
uritemplate==3.0.0
url-normalize==1.3.3
urllib3==1.22
versioneer==0.18
